Add ClearBtn file
