+++ src/App.js
@@ -1,23 +1,7 @@
 import React, { Component } from "react";
 import NumBtn from "./components/NumBtn";
 import ExecutorBtn from "./components/ExecutorBtn";
-
-class ClearBtn extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <button
-        className="col mr-1 btn btn-success"
-        onClick={() => this.props.onClick()}
-      >
-        {this.props.cmd}
-      </button>
-    )
-  }
-}
+import ClearBtn from "./components/ClearBtn";
 
 class MemoryBtn extends Component {
   constructor(props) {
